+++ eslint.backend.config.mjs
@@ -1,0 +1,24 @@
+import eslint from "@eslint/js";
+import globals from "globals";
+
+export default [
+  {
+    ignores: ["node_modules/"],
+  },
+  {
+    ...eslint.configs.recommended,
+    languageOptions: {
+      globals: {
+        ...globals.node,
+      },
+    },
+  },
+  {
+    files: ["**/public/**/*.js"],
+    languageOptions: {
+      globals: {
+        ...globals.browser,
+      },
+    },
+  },
+];